# speedup BitSet iteration (by 2x or 3x)