Cf. https://discourse.julialang.org/t/iterate-and-intersect-on-bitset-vs-int/30345/5 for motivation.

A shorter (stupid) benchmark:
```julia
b1 = BitSet(1:1000)
b2 = BitSet(rand(1:1000, 500))
b3 = BitSet(rand(1:1000, 100))

function findval(set, x)
    for v in set
        if v == x
            return true
        end
    end
    false
end
```
Master:
```julia
julia> for b = (b1, b2, b3)
           @btime findval($b, 1001)
       end
  6.939 μs (0 allocations: 0 bytes)
  3.025 μs (0 allocations: 0 bytes)
  660.325 ns (0 allocations: 0 bytes)
```
PR:
```julia
julia> for b = (b1, b2, b3)
           @btime findval($b, 1001)
       end
  2.602 μs (0 allocations: 0 bytes)
  1.140 μs (0 allocations: 0 bytes)
  225.121 ns (0 allocations: 0 bytes)
```
EDIT: PR with `_blsr`:
```julia
julia> for b = (b1, b2, b3)
           @btime findval($b, 1001)
       end
  1.071 μs (0 allocations: 0 bytes)
  499.384 ns (0 allocations: 0 bytes)
  110.638 ns (0 allocations: 0 bytes)
```

~~This is a speedup of about 2.6x for `b1` and `b2`, and almost 3x for `b3`.~~
This is a speedup of about 6x.