@chethega 2019-10-27
Consider using the same algorithm as logical indexing, for a further ~40% speedup:
```
julia> @eval Base function iterate(s::BitSet, (word, idx) = (CHK0, 0))
       while (word==0)
           idx == length(s.bits) && return nothing
           idx += 1
           word = @inbounds s.bits[idx]
           end
           ret = (idx -1 + s.offset)<<6 + trailing_zeros(word)
           ret, (_blsr(word), idx)
       end
```
Reason for this is that `TZCNT` has high latency compared to `BLSR`, so we want it out of the dependency chain. This is better for superscalar execution.