@rfourquet 2019-10-28
Wow amazing! Actually, I thought that `trailing_zero(word) == 64` might be more expensive that `word == 0`, but just putting it out of the loop was not enough to get better performance. 

So in your example you still compute `b = trailing_zeros(word)`. I would then naïvely believe that the difference is that `_blsr(word)` is way faster than `xor(word, one(UInt64) << b)`... but do I understood corretly from your explanation that actually the difference comes from that in your version, `trainling_zeros` is used only for the returned value `ret`, and not for the returned state, which means that the CPU can start doing stuff for the next iteration even before having computed `ret`? Sorry, fuzzy words but I'm not familiar with CPUs :D

> Consider using the same algorithm as logical indexing

For those wondering, this is the `findall(::BitArray)` method, which I hadn't realized does indeed basically the same thing as `BitSet` iteration. And @chethega implemented a fast algorithn with this `_blsr` trick in #29746.