@mbauman 2019-10-28
Looks like we do have some BitSet perf tests in the `collection` group.

@nanosoldier `runbenchmarks("collection", vs = ":master")`