@nanosoldier 2019-10-28
[Your benchmark job](https://github.com/JuliaLang/julia/pull/33690#issuecomment-546977902) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/d68a31d50fd34cbbd83443563b674ddec87df09c/d9ecc63_vs_88c34fc/report.md). cc @ararslan