@vchuravy 2019-10-28
Would be good to add a benchmark for this.